feature(Coinify): separate quote input limits and error text out into its own component
